backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-core
langchain-mistralai
langchain-google-genai
fastapi
uvicorn